Add new site structure and theme configuration
This commit introduces a new site structure and theme configurations. The new Jekyll theme, Minima, has been configured as per the needs of the project. This includes layout settings and design skin for the home layout, along with necessary plugins. Three new pages - index, about and 404 were created. 'About' and 'index' pages contain a brief about BEE6 - a Discord bot, while the '404' page is a custom page to handle page not found errors. The README file was also updated to serve as a homepage for the site.